+++ Assets/UtilityScripts/com.dman.utilities/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [0.3.0] - 2021-01-16
+
+mouse over raycast helper. remove isMouseDown check on the RaycastToObject method.
+
+### Added
+
+- RaycastGroup to cache raycasts into the environment
+
 ## [0.2.0] - 2021-01-16
 
 serializable vectors.